Passed 'mode' argument to LangChain wrapper
Apologies if this was excluded for a reason I am not understanding. Lance and the LanceDataset VectorStore are the most clean and effective method I've found for fast I/O for local chat bots. Allowing appending will let me to work with just the VectorStore implementation thanks to the save/local_load methods (which are conspicuously missing from other VecStore implementations). I hope to see the Lance pull request accepted soon, but in the interim thought I would share so as to not have to rely on my own local versions when updating packages, or in the worst case, be told how I should actually be doing this if not the way I've proposed.